+++ src/blueprint/operators/const_field.rs
@@ -6,7 +6,6 @@ use crate::config;
 use crate::config::Field;
 use crate::lambda::Expression;
 use crate::lambda::Expression::Literal;
-use crate::mustache::Mustache;
 use crate::try_fold::TryFold;
 use crate::valid::{Valid, ValidationError, Validator};
 
@@ -29,21 +28,6 @@ pub struct CompileConst<'a> {
     pub field: &'a config::Field,
     pub value: &'a DynamicValue,
     pub validate: bool,
-}
-
-#[derive(Debug, Clone)]
-pub enum MustacheOrValue {
-    Mustache(Mustache),
-    Value(serde_json::Value),
-}
-
-impl MustacheOrValue {
-    pub fn is_const(&self) -> bool {
-        match self {
-            MustacheOrValue::Mustache(m) => m.is_const(),
-            _ => true,
-        }
-    }
 }
 
 pub fn compile_const(inputs: CompileConst) -> Valid<Expression, String> {
@@ -94,8 +78,7 @@ pub fn compile_const(inputs: CompileConst) -> Valid<Expression, String> {
     }
 }
 
-pub fn update_const_field<'a>(
-) -> TryFold<'a, (&'a ConfigModule, &'a Field, &'a config::Type, &'a str), FieldDefinition, String>
+pub fn update_const_field<'a>() -> TryFold<'a, (&'a ConfigModule, &'a Field, &'a config::Type, &'a str), FieldDefinition, String>
 {
     TryFold::<(&ConfigModule, &Field, &config::Type, &str), FieldDefinition, String>::new(
         |(config_module, field, _, _), b_field| {
@@ -107,10 +90,10 @@ pub fn update_const_field<'a>(
                 DynamicValue::try_from(&const_field.data.clone())
                     .map_err(|e| ValidationError::new(e.to_string())),
             )
-            .and_then(|value| {
-                compile_const(CompileConst { config_module, field, value: &value, validate: true })
-                    .map(|resolver| b_field.resolver(Some(resolver)))
-            })
+                .and_then(|value| {
+                    compile_const(CompileConst { config_module, field, value: &value, validate: true })
+                        .map(|resolver| b_field.resolver(Some(resolver)))
+                })
         },
     )
 }
